+++ CotEditor/Sources/Settings Window/Panes/FormatSettingsView.swift
@@ -312,9 +312,12 @@ private struct SyntaxListView: View {
                         } catch let error as ImportDuplicationError {
                             self.importingError = error
                             self.isImportConfirmationPresented = true
+                            return
                         } catch {
                             self.error = error
+                            return
                         }
+                        self.selection = self.manager.state(of: name)
                     }
                 case .failure(let error):
                     self.error = error