+++ CotEditor/Sources/Text Finder/Multiple Replace/Views/MultipleReplaceListView.swift
@@ -120,9 +120,12 @@ struct MultipleReplaceListView: View {
                         } catch let error as ImportDuplicationError {
                             self.importingError = error
                             self.isImportConfirmationPresented = true
+                            return
                         } catch {
                             self.error = error
+                            return
                         }
+                        self.selection = name
                     }
                 case .failure(let error):
                     self.error = error