+++ CHANGELOG.md
@@ -22,7 +22,7 @@
 
 ### Known Issues
 
-- The text added to CotEditor 6.2.0 are not yet localized to Czech.
+- The text added to CotEditor 6.2.0 are not fully localized yet.